Update README.md
Happiness and excitement, a sense of achievement, and a sense of connection to collaborate with friends in a black team. These are things I have gained from esports,of course,it's also true in basketball and I hope to have the opportunity to continue to experience them in the future.